movement
words cannot describe how much of a struggle this was to implement, but. have implemented an ephemeral memory for the LLM's further testing still yet to be done, but within your conversation, you can bet the LLM will ahve an inkling of what went on, this took me maybe a total of 12-15 hours of my time? had to refactor the base models, add a new class for handling conversations, just making stuff more object oriented rather than functional, im thinking i might have to move forward with the way i code like this. its pretty nice and fun to do.