movement
made some changes to how things are called, so now the search function is now a class called search tool, this provides the ability to eventually introduce plug ins for users, following a way of inheritance, they can create a plug in / tool, dictate what it does e.g. access an api, edit a document from an api, maybe just gather data from that api, i think there'll be more use cases, but for now it works, and thats all that matters :).